+++ sbi-rt/src/base.rs
@@ -2,10 +2,9 @@
 
 use crate::binary::{sbi_call_0, sbi_call_1};
 
-pub use sbi_spec::base::Version;
 use sbi_spec::base::{
-    EID_BASE, GET_MARCHID, GET_MIMPID, GET_MVENDORID, GET_SBI_IMPL_ID, GET_SBI_IMPL_VERSION,
-    GET_SBI_SPEC_VERSION, PROBE_EXTENSION,
+    Version, EID_BASE, GET_MARCHID, GET_MIMPID, GET_MVENDORID, GET_SBI_IMPL_ID,
+    GET_SBI_IMPL_VERSION, GET_SBI_SPEC_VERSION, PROBE_EXTENSION,
 };
 
 /// Return the current SBI specification version.